Day 9.2 - Scrapped the optimization and it works. The real optimization was the friends we made alo... no, actually it was changing the LinkedList to an ArrayList.

+++ src/main/java/org/tot/aoc/Day9.java
@@ -1,15 +1,13 @@
 package org.tot.aoc;
 
-import org.apache.commons.lang3.Range;
-
 import java.nio.CharBuffer;
 import java.util.*;
-import java.util.function.Predicate;
 
 public class Day9 {
 
-    List<FileBlock> filesystem = new LinkedList<>();
-    SortedMap<Integer, Integer> freeSpaceMarkers = new TreeMap<>();
+    List<FileBlock> filesystem = new ArrayList<>();
+
+    int seekRangeStart = 0;
 
     public long solvePuzzle1(String input) {
 
@@ -68,21 +66,18 @@ public class Day9 {
             filesystem.add(new FileBlock(maxFileId, size, free));
         }
 
-        int seekRangeStart = 0;
         int seekRangeEnd = filesystem.size() - 1;
 
 
         System.out.println("Defragging...");
-//        printFilesystem();
 
-        nextFieldId:
         for (long fileId = maxFileId; fileId > 0; fileId--) {
 
             if (fileId % 100 == 0) {
                 System.out.println(fileId);
             }
 
-            int fromIndex = indexOfReverse(filesystem, seekRangeEnd, seekRangeStart, fileId);
+            int fromIndex = indexOfReverse(filesystem, seekRangeEnd, fileId);
             if (fromIndex < 0) {
                 throw new IllegalStateException("FileBlock not found " + fileId);
             }
@@ -94,39 +89,15 @@ public class Day9 {
 
             int freeSpaceIdx = indexOfFreeSpace(moving.size, fromIndex);
             if (freeSpaceIdx < 0) {
-//                System.out.println("Not moving, no space " + fileId);
-//                printFilesystem();
                 continue;
             }
 
             relocate(moving, fromIndex, freeSpaceIdx);
-//            printFilesystem();
-
-//            FileBlock gap = filesystem.get(fromIndex - 1);
-//
-//            for (int i = seekRangeStart; i < seekRangeEnd; i++) {
-//                FileBlock preceding = filesystem.get(i);
-//                if (preceding.freeSpaceAfter >= moving.size) {
-//                    gap.freeSpaceAfter += moving.size + moving.freeSpaceAfter;
-//                    moving.freeSpaceAfter = preceding.freeSpaceAfter - moving.size;
-//                    preceding.freeSpaceAfter = 0;
-//
-//                    filesystem.remove(fromIndex);
-//                    filesystem.add(i + 1, moving);
-//                    System.out.printf("Moved %d to %d%n", fileId, i + 1);
-//                    continue nextFieldId;
-//                }
-//
-//            }
-//            System.out.println("Not moving, no space " + fileId);
-
-
         }
 
         long checksum = 0;
         int idx = 0;
         for (var f : filesystem) {
-//            System.out.printf("%d %d %n", f.fileId, idx);
             int end = idx + f.size;
             while (idx < end) {
                 checksum += (idx * f.fileId);
@@ -138,8 +109,8 @@ public class Day9 {
         return checksum;
     }
 
-    int indexOfReverse(List<FileBlock> list, int start, int end, long findFileId) {
-        for (int j = start; j >= end; j--) {
+    int indexOfReverse(List<FileBlock> list, int start, long findFileId) {
+        for (int j = start; j >= 0; j--) {
             if (list.get(j).fileId == findFileId) {
                 return j;
             }
@@ -149,36 +120,14 @@ public class Day9 {
 
     int indexOfFreeSpace(int size, int notBeyond) {
 
-        var itor = freeSpaceMarkers.tailMap(size).entrySet().iterator();
-        while (itor.hasNext()) {
-            var entry = itor.next();
-            if (entry.getKey() >= size) {
-                int index = entry.getValue();
-                if (index < notBeyond) {
-                    itor.remove();
-                    return index;
-                }
-            }
-        }
-
-
-        int start = freeSpaceMarkers.values()
-                .stream()
-                .sorted()
-                .findFirst()
-                .orElse(0);
-//        System.out.printf("Cache miss for %d, starting at %d %n", size, start);
-
-        for (int i = start; i < notBeyond; i++) {
+        for (int i = 0; i < notBeyond; i++) {
             FileBlock b = filesystem.get(i);
             int freeSpace = b.freeSpaceAfter;
-            if (freeSpace < size) {
-                Integer existingIndex = freeSpaceMarkers.getOrDefault(freeSpace, Integer.MAX_VALUE);
-                if (i < existingIndex) {
-                    freeSpaceMarkers.put(freeSpace, i);
-                }
+            if (freeSpace == 1) {
+                seekRangeStart = i;
+            }
 
-            } else {
+            if (freeSpace >= size) {
                 return i;
             }
         }
@@ -203,6 +152,11 @@ public class Day9 {
 
         fromNeighbor.freeSpaceAfter += moveFile.size + moveFile.freeSpaceAfter;
         moveFile.freeSpaceAfter = toNeighbor.freeSpaceAfter - moveFile.size;
+
+        if (moveFile.freeSpaceAfter < 0) {
+            throw new IllegalStateException();
+        }
+
         toNeighbor.freeSpaceAfter = 0;
 
         filesystem.remove(from);
